refactor: stopwatch page
