+++ src/pages/mission/[id]/stopwatch/TextSection.tsx
